Performance: BootReceiver: set CPU min/max freqs before CPU governor
When setting certain combinations of CPU governor/freqs on boot , the
max frequency needs to be changed first in order for some settings to
work correctly.

Example: changing from a default governor of ondemand and frequency
of 1GHz to interactive governor and 1.2GHz frequency , hispeed_freq
will get stuck at 1GHz unless the max frequency is set prior to the
governor.