Settings: EditTextPreference instead of ListPreference for SMS limit
This changes the preference for how many SMS's can be sent before an AlertDialog is shown from a ListPreference to an EditTextPreference.
The  current limited options are 30, 50, 100 & 200 messages per 15 minutes. A few people send more than the current highest limit of 200 while I'm on the other end of the spectrum and prefer a low limit (I had mine set on 15 on CM10).
With this patch, the user can set the limit to whatever they want.